💡 chore    : views 파일명 변경
